+++ lib/src/internal/tabs_area/tabs_area_layout_parent_data.dart
@@ -1,16 +1,19 @@
 import 'package:flutter/rendering.dart';
 import 'package:meta/meta.dart';
+import 'package:tabbed_view/src/internal/size_holder.dart';
 
 /// Parent data for [_TabsAreaLayoutRenderBox] class.
 @internal
 class TabsAreaLayoutParentData extends ContainerBoxParentData<RenderBox> {
   bool visible = true;
   bool selected = false;
+  SizeHolder? sizeHolder;
 
   /// Resets all values.
   void reset() {
     visible = true;
     selected = false;
+    sizeHolder?.size = null;
   }
 }